[rocjitsu] DBIT: Moved shared units out of dbt/ (NBC)
An upcoming DBI PR found some dbt/ code helpful. This moves that
code:
* code/dbt/scoped_cfg_edges -> code/
* tests/dbt/support/translate_test_support ->
  tests/amdgpu_elf_test_support
* tests/dbt/support/elf_test_support -> tests/

Moves and related updates ONLY. No behavior change.
