+++ emulation/rocjitsu/lib/rocjitsu/src/rocjitsu/code/scoped_cfg_edges.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 Advanced Micro Devices, Inc.
 // SPDX-License-Identifier: MIT
 
-#include "rocjitsu/code/dbt/scoped_cfg_edges.h"
+#include "rocjitsu/code/scoped_cfg_edges.h"
 
 #include "rocjitsu/code/basic_block.h"
 #include "rocjitsu/isa/instruction.h"
